# openapi.json
{"openapi":"3.1.0","info":{"title":"IAM Microservice","version":"0.1.0"},"paths":{"/scim/v2/Users":{"post":{"tags":["SCIM"],"summary":"Create User","operationId":"create_user_scim_v2_Users_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SCIMUserIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["SCIM"],"summary":"Search Users","operationId":"search_users_scim_v2_Users_get","parameters":[{"name":"filter","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filter"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/scim/v2/Users/{uid}":{"get":{"tags":["SCIM"],"summary":"Get User","operationId":"get_user_scim_v2_Users__uid__get","parameters":[{"name":"uid","in":"path","required":true,"schema":{"type":"string","title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["SCIM"],"summary":"Patch User","operationId":"patch_user_scim_v2_Users__uid__patch","parameters":[{"name":"uid","in":"path","required":true,"schema":{"type":"string","title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["Auth"],"summary":"Token","operationId":"token_auth_token_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Auth"],"summary":"Me","operationId":"me_auth_me_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/authz/evaluate":{"post":{"tags":["ABAC"],"summary":"Evaluate","operationId":"evaluate_authz_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SCIMEmail":{"properties":{"value":{"type":"string","title":"Value"},"primary":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Primary","default":false}},"type":"object","required":["value"],"title":"SCIMEmail"},"SCIMName":{"properties":{"givenName":{"type":"string","title":"Givenname"},"familyName":{"type":"string","title":"Familyname"}},"type":"object","required":["givenName","familyName"],"title":"SCIMName"},"SCIMUserIn":{"properties":{"userName":{"type":"string","title":"Username"},"name":{"$ref":"#/components/schemas/SCIMName"},"active":{"type":"boolean","title":"Active","default":true},"emails":{"items":{"$ref":"#/components/schemas/SCIMEmail"},"type":"array","title":"Emails"},"groups":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Groups"}},"additionalProperties":true,"type":"object","required":["userName","name","emails"],"title":"SCIMUserIn"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}